Don't remount if the container is already running
This change fixes double-mounting when running an exec or a cp on a
running container with the containerd integration enabled. Sharing
layers with overlayfs results in an undefined state and things start to
randomly break with a "No such file or directory" error.

Graph drivers use reference counting and only mount things once, we
can't do this with containerd because we don't know where it mounted the
container rootfs. Instead of reference counting we return the
`/proc/<pid>/root` directory for exec and/or cp to read from.

Signed-off-by: Djordje Lukic <djordje.lukic@docker.com>