Allow for xattr copy failure for vfs
vfs is declared to work with any filesystem, but after
https://github.com/moby/moby/commit/31f654a704f61768828d5950a13f30bb493d1239
it's no longer working with NFS.

As the extended attribute support depends on filesystem and
if we do copy it in vfs and do not allow failure, that would
essentially mean that vfs does NOT support all filesystems but
only those that support xattr.

So we should just try to copy security.capabilities and allow
for failure. In this way, vfs come back to the state of
being able to run on any filesystem as declared in
https://docs.docker.com/storage/storagedriver/select-storage-driver/.

Fixes https://github.com/moby/moby/issues/45417

Signed-off-by: Chen Qi <Qi.Chen@windriver.com>