libcontainerd: work around exec start bug in c8d
It turns out that the unnecessary serialization removed in
b75246202ab9b1e5bb94c377f90db8ed38cfa0e0 happened to work around a bug
in containerd. When many exec processes are started concurrently in the
same containerd task, it takes seconds to minutes for them all to start.
Add the workaround back in, only deliberately this time.

Signed-off-by: Cory Snider <csnider@mirantis.com>