builder/worker: Handle image exporter same as moby
Moby exporter is a wrapper for the `image` exporter that modifies the
request attributes on the fly to make it work correctly with the engine.

Before this, if user explicitly asks for the `image` exporter, it would
completely bypass the special handling it needs. For example it wouldn't
create an image if the user didn't provide a tag.

This makes the `image` exporter being handled the same way as `moby.

Signed-off-by: Paweł Gronowski <pawel.gronowski@docker.com>